Make sure we consistently handle checking unit strings to identify descriptor type
We were identifying unit values `CCITT IA5` and `CCITT_IA5` as
indicating descriptor type string, but not `CCITTIA5`. This was causing
some tests to fail with the new BUFR files added to thredds-test-data. Now
we do this check in one method. This PR also makes sure we do the same for
Code Table and Flag table units (also checking for `CODETABLE` and
`FLAGTABLE`, as that is what is used in the eumetsat tables). Oh,
BUFR...you so silly.